requirements.txt
pytest==8.3.3
python-dotenv==1.0.1
setuptools==75.2.0
spotipy==2.24.0
